changed ProtocolIO Sstruct to accept 64bit I/O

+++ src/multiplication.h
@@ -1,6 +1,6 @@
 void oblivMul(void* args);
 
 typedef struct {
-	int data;
-	int result;
+	long long data;
+	long long result;
 } ProtocolIO;
